added wikipedia search function
